added quiz and question answer

+++ src/main/java/com/example/demo/repository/LessonRepository.java
@@ -12,5 +12,5 @@ public interface LessonRepository extends JpaRepository<Lesson, Integer>{
 	List<Lesson> findByModuleModuleIdOrderByPosition(Integer moduleId);
 	
 	List<Lesson> findByModuleModuleIdAndFreePreviewTrue(Integer moduleId);
-	int countByModuleModuleId(Integer moduleId);
+	long countByModuleModuleId(Integer moduleId);
 }
